Feat: Modification du style de la timeline
Pour mettre le trait de la timeline sur tous les éléments sauf le dernier
